ai_assistant/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic-settings
sqlalchemy
alembic
pymysql
python-multipart
httpx
loguru
python-jose[cryptography]
passlib[bcrypt]
boto3
openai
pytest
pytest-asyncio
ruff
black